requirements.txt
clip==0.2.0
nltk==3.6.5
numpy==1.22.3
Pillow==9.1.1
requests==2.27.1
sentence_transformers==2.2.0
torch==1.12.0.dev20220411
tqdm==4.62.3
transformers==4.11.3